Ensure Access Controls on link resource
When a link resource needs to change its destination, there was a bug in
the logic flow that would cause the access controls on the resulting
symlink to revert to the default access controls. This commit closes
that hole by re-evaluating the current access controls after creating
the new symlink.

Signed-off-by: Tom Duffield <tom@chef.io>